Make TransferError::Unknown hold u32
macOS errors have the high bit set

+++ src/platform/linux_usbfs/mod.rs
@@ -31,6 +31,6 @@ fn errno_to_transfer_error(e: Errno) -> TransferError {
         Errno::PROTO | Errno::ILSEQ | Errno::OVERFLOW | Errno::COMM | Errno::TIME => {
             TransferError::Fault
         }
-        _ => TransferError::Unknown(e.raw_os_error()),
+        _ => TransferError::Unknown(e.raw_os_error() as u32),
     }
 }
